[FIX] stock: onchange performance on picking
The use case is similar to a297d549af9e086d5e1c530c99c76afca061b37d, but
applied to pickings.

When the picking contains many stock move, the manual edition of the
quantitites is quite slow due to the `onchange` call. In the example of
the OPW with 300 stock moves, the modification of the done quantity
triggers an `onchange` which lasts about 1 second.

The reason is the same than in the mentioned commit: the `onchange` is
triggered on all lines in the case of an inline tree view, and there is
not much we can do to optimize the business logic.

Therefore, the solution is the same: add a stat button to access the
stock moves directly in a tree view instead, to avoid this `onchange`.

opw-1965248

closes odoo/odoo#33004

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>