[FIX] website_forum: fix translation from faq
Since the refactoring of forum, faq sanitize is set to False.
The editor of the faq can make what he want...

But disable sanitize will break the translation since fields.py has code:
```python
     def _setup_attrs(self, model, name):
        super(Html, self)._setup_attrs(model, name)
        # Translated sanitized html fields must use html_translate or a callable.
        if self.translate is True and self.sanitize:
            self.translate = html_translate
```

Now we don't become html_translate magically, but we need to specify it
in the field definition direclty.

Courtesy of @mart-e

This commit closes opw-1978662

closes odoo/odoo#33112

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>