[IMP] website_sale: change product url to include variants
Hash containing variants unique id added to the url.
In the shop when the product page is loaded, the variants are now
selected using this hash.
The product.product website_url is no longer the template's url, it is
now using the hash and will redirect you to the right product.

task-1958091