[FIX] website_sale: clean the action to send recovery cart
The action to send cart recovery mails is really made to be automated.
However it was written in a way requiring either manual input, to use
action_recovery_email_send.
To make an automated action, you basically need to put the code of it
directly into the action itself, or it would be made incorrectly.

opw 1981665

closes odoo/odoo#33185

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>