[FIX] mrp_byproduct: Impossible to process a work order
Steps to reproduce the bug:

- Create two storable products A and B
- Create a BOM for A with B as component and A as by product
- Set this BOM as "To manufacture"
- Create a MO for A and try to process it

Bug:

Ensure one error was raised by function check_finished_move_lots defined in module mrp
because the variable produce_move must be one record.

Now, it's not possible to create a MO for A if A has a BOM with A as by product.

opw:1974863

closes odoo/odoo#33280

Signed-off-by: Simon Goffin (sig) <sig@openerp.com>